+++ Tyrant/Assets/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 322133891b34d4cd18e9ea6ddfe015dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
