+++ Tyrant/Assets/Animation/Controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36176df350da64ca7a32f1dd9ec9f242
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
