+++ Tyrant/Assets/Animation/Controller/chort_idle_anim_f0.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a7da0aeb96a241519b28dd1e2f6be11
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
