+++ Tyrant/Assets/Animation/Enemy Anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 63999c473c6b244aca4b3351afc155e3
+guid: ac85ae00360f1456aabefe33f6f52f6d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
