+++ Tyrant/Assets/Animation/Enemy Anim/chort.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9a11e861a5c0481f8cda14696979579
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
