+++ Tyrant/Assets/Animation/Enemy Anim/chort/idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6e30be277faa43c8b374f0b0aa0e988
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
