+++ Tyrant/Assets/Animation/Enemy Anim/chort/run.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b3136b5e164fc4417a135bf4a66c137c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
