--- Tyrant/Assets/Chort.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ce804d4582324c419607e82a232d087
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
