+++ Tyrant/Assets/Prefabs/Enemy/Enemy.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d56aa9516e1109a42b534066b396c81c
+guid: 8b9f0df936cf54392b9c2877c62446d7
 PrefabImporter:
   externalObjects: {}
   userData:
