--- Tyrant/Assets/Prefabs/Enemy/NormalEnemy.prefab
@@ -1,71 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &3573339058323897713
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3722758443527116800, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: enemyState.enemyDamage
-      value: 10
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116800, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: enemyState.enemyHealth
-      value: 50
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116800, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: enemyState.enemyMoveSpeed
-      value: 2
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116801, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_Name
-      value: NormalEnemy
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3722758443527116802, guid: 65db439a9128a684a878a176e76edae6, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 65db439a9128a684a878a176e76edae6, type: 3}
