+++ Tyrant/Assets/Prefabs/Enemy/normal enemy.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65db439a9128a684a878a176e76edae6
+guid: 96eef6fb1157b485d86e8003cc10e7f1
 PrefabImporter:
   externalObjects: {}
   userData:
