+++ Tyrant/Assets/Scripts/Enemy/Enemy.cs
@@ -17,13 +17,22 @@ public class Enemy : MonoBehaviour
    
     public EnemyState EnemyState { get => enemyState; set => enemyState = value; }
 
+   
 
-    // Start is called before the first frame update
-    void Start()
+
+    /// <summary>
+
+    private Animator anim;
+/// </summary>
+
+
+// Start is called before the first frame update
+void Start()
     {
         behaviours = gameObject.GetComponent<StaticMachine>();
         behaviours.setEnemy(this);
         behaviours.AllBehaviour();
+        anim = GetComponent<Animator>();
     }
 
     // Update is called once per frame
@@ -31,16 +40,58 @@ public class Enemy : MonoBehaviour
     {
         behaviours.Update();
         distance = Vector3.Distance(transform.position, target.position);
-        if (distance < attatckdistance)
+        if (Vector3.Distance(transform.position, target.position) > enemyState.StopDistance)
         {
             enemyState.force = behaviours.ForceCalculate();
             enemyState.acceleration = enemyState.force / enemyState.Mass;
             enemyState.velocity += enemyState.acceleration;
+
+            //animation
+            anim.SetBool("isRunning", true);
         }
         else
         {
             enemyState. velocity = Vector3.zero;
+/*
+            if (Time.time >= enemyState.TimeBetweenAttacks)
+            {
+
+                enemyState.TimeBetweenAttacks = Time.time + enemyState.TimeBetweenAttacks;
+            }
+*/
+            //animation
+            anim.SetBool("isRunning", false);
         }
          transform.position += enemyState.velocity;
+
+
     }
+    /*------------------攻击位移------------------------
+    IEnumerator Attack()
+    {
+        //player.GetComponent<Player>().TakeDamage(damage);
+
+        Vector2 originalPosition = transform.position;
+
+        Vector2 targetPosition = player.position;
+
+        float percent = 0;
+        while (percent <= 1)
+        {
+            percent += Time.deltaTime * attackSpeed;
+
+            float formula = (-Mathf.Pow(percent, 2) + percent) * 4;
+
+            transform.position = Vector2.Lerp(originalPosition, targetPosition, formula);
+
+            yield return null;
+        }
+    }
+    */
+    void Animation()
+    {
+        anim.SetBool("isRunning", true);
+    
+    }
+
 }
