+++ Tyrant/Assets/Scripts/Enemy/EnemyState.cs
@@ -13,7 +13,12 @@ public class EnemyState
     private float enemyMoveSpeed = 0f;      // default value
     [SerializeField]
     private float mass = 1f;
-
+    [SerializeField]
+    private float timeBetweenAttacks =0f;
+    [SerializeField]
+    private float stopDistance = 0f;
+    [SerializeField]
+    private float attackSpeed = 0f;
 
     public Vector3 velocity { get; set; }
     public Vector3 acceleration { get; set; }
@@ -23,4 +28,7 @@ public class EnemyState
     public float EnemyDamage { get => enemyDamage; }
     public float EnemyMoveSpeed { get => enemyMoveSpeed; }
     public float Mass { get => mass; }
+    public float TimeBetweenAttacks { get => timeBetweenAttacks; set =>timeBetweenAttacks= value; }
+    public float StopDistance { get => stopDistance; }
+    public float AttackSpeed { get => attackSpeed; }
 }
