+++ Tyrant/UserSettings/EditorUserSettings.asset
@@ -11,6 +11,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c78111b122b292926237e38271427fb
       flags: 0
+    RecentlyUsedScenePath-2:
+      value: 22424703114646680e0b0227036c7a1e131a016439262f2434
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
